+++ indico/web/client/js/jquery/widgets/jinja/ckeditor_widget.js
@@ -11,9 +11,12 @@ import {getConfig} from 'indico/ckeditor';
 
 (function(global) {
   global.setupCKEditorWidget = async function setupCKEditorWidget(options) {
-    const {fieldId, images = true, width, height = 475, ...rest} = options;
+    const {fieldId, images = true, imageUploadURL = null, width, height = 475, ...rest} = options;
     const field = document.getElementById(fieldId);
-    const editor = await ClassicEditor.create(field, {...getConfig({images}), ...rest});
+    const editor = await ClassicEditor.create(field, {
+      ...getConfig({images, imageUploadURL}),
+      ...rest,
+    });
     editor.setData(field.value);
     editor.model.document.on('change:data', () => {
       field.value = editor.getData();